requirements.txt
aiohttp
aiofiles
mojang-api-3
discord.py